# docs/swagger/swagger.yaml
definitions:
  ActiveSubstanceDto:
    properties:
      concentration:
        format: float32
        type: number
      id:
        type: string
    type: object
  ActiveSubstanceResDto:
    properties:
      concentration:
        type: number
      id:
        type: string
    type: object
  AddDosageRuleDto:
    properties:
      dosage:
        $ref: '#/definitions/DosageRuleDto'
      medicine_id:
        type: string
    required:
    - dosage
    - medicine_id
    type: object
...
  AddMedicineDto:
    properties:
      contraindications:
        items:
          type: string
        type: array
      dosages:
        items:
          $ref: '#/definitions/DosageRuleDto'
        type: array
      effect_on_driver:
        type: boolean
      effect_on_pregnant:
        type: boolean
      expire_time:
        type: integer
      form_id:
        type: string
      is_prescription:
        type: boolean
      method_of_application:
        type: string
      name:
        type: string
      recommendations:
        items:
          type: string
        type: array
      substances:
        items:
          $ref: '#/definitions/ActiveSubstanceDto'
        type: array
      unit_id:
        type: string
    required:
    - contraindications
    - dosages
    - effect_on_driver
    - effect_on_pregnant
    - expire_time
    - form_id
    - is_prescription
    - method_of_application
    - name
    - recommendations
    - substances
    - unit_id
    type: object
  AdminResDto:
    properties:
      contraindications:
        items:
          type: string
        type: array
      dosages:
        items:
          $ref: '#/definitions/DosageRuleResDto'
        type: array
      effect_on_driver:
        type: boolean
      effect_on_pregnant:
        type: boolean
      expire_time:
        type: integer
      form_id:
        type: string
      id:
        type: string
      is_prescription:
        type: boolean
      method_of_application:
        type: string
      name:
        type: string
      recommendations:
        items:
          type: string
        type: array
      substances:
        items:
          $ref: '#/definitions/ActiveSubstanceResDto'
        type: array
      unit_id:
        type: string
    type: object
  AppError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  DosageRuleDto:
    properties:
      dosageValue:
        format: float32
        type: number
      numberOfDosesPerDay:
        type: integer
      type:
        $ref: '#/definitions/DosageType'
      valueFrom:
        type: integer
      valueTo:
        type: integer
    type: object
  DosageRuleItemDto:
    properties:
      id:
        type: string
      label:
        type: string
    type: object
  DosageRuleResDto:
    properties:
      dosage_value:
        type: number
      id:
        type: string
      number_of_doses_per_day:
        type: integer
      type:
        $ref: '#/definitions/DosageTypeResDto'
      value_from:
        type: integer
      value_to:
        type: integer
    type: object
  DosageType:
    enum:
    - weight
    - age
    type: string
    x-enum-varnames:
    - ByWeight
    - ByAge
  DosageTypeResDto:
    enum:
    - weight
    - age
    type: string
    x-enum-varnames:
    - ByWeight
    - ByAge
...
  RemoveDosageRuleDto:
    properties:
      rule_id:
        type: string
    required:
    - rule_id
    type: object
...
  RemoveMedicineDto:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  SuccessResDTO:
    properties:
      status:
        type: string
    type: object
  UpdateCompositionDto:
    properties:
      medicine_id:
        type: string
      substances:
        items:
          $ref: '#/definitions/ActiveSubstanceDto'
        type: array
    required:
    - medicine_id
    - substances
    type: object
  UpdateLinksDto:
    properties:
      ids:
        items:
          type: string
        type: array
      medicine_id:
        type: string
    required:
    - ids
    - medicine_id
    type: object
  UpdateMedicineDto:
    properties:
      effect_on_driver:
        type: boolean
      effect_on_pregnant:
        type: boolean
      expire_time:
        type: integer
      form_id:
        type: string
      id:
        type: string
      is_prescription:
        type: boolean
      method_of_application:
        type: string
      unit_id:
        type: string
    required:
    - effect_on_driver
    - effect_on_pregnant
    - expire_time
    - form_id
    - id
    - is_prescription
    - method_of_application
    - unit_id
    type: object
...
info:
  contact: {}
paths:
  /admin/dosage-rules:
    get:
      description: Возвращает правила дозировок с названиями лекарств
      produces:
      - application/json
      responses:
        "200":
          description: Список правил
          schema:
            items:
              $ref: '#/definitions/DosageRuleItemDto'
            type: array
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Список правил дозировки
      tags:
      - Admin
  /admin/medicine:
    delete:
      consumes:
      - application/json
      description: Удаляет лекарство по идентификатору
      parameters:
      - description: Идентификатор лекарства
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RemoveMedicineDto'
      produces:
      - application/json
      responses:
        "200":
          description: Лекарство удалено
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Удаление лекарства
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: Обновляет данные лекарства
      parameters:
      - description: Поля для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateMedicineDto'
      produces:
      - application/json
      responses:
        "200":
          description: Лекарство обновлено
          schema:
            $ref: '#/definitions/AdminResDto'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
        "404":
          description: Лекарство не найдено
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Обновление лекарства
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Создает новое лекарство
      parameters:
      - description: Данные лекарства
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/AddMedicineDto'
      produces:
      - application/json
      responses:
        "201":
          description: Лекарство создано
          schema:
            $ref: '#/definitions/AdminResDto'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Добавление лекарства
      tags:
      - Admin
  /admin/medicine/composition:
    patch:
      consumes:
      - application/json
      description: Обновляет список действующих веществ
      parameters:
      - description: Состав лекарства
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateCompositionDto'
      produces:
      - application/json
      responses:
        "200":
          description: Состав обновлен
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Обновление состава
      tags:
      - Admin
  /admin/medicine/contraindications:
    patch:
      consumes:
      - application/json
      description: Обновляет список противопоказаний для лекарства
      parameters:
      - description: Список противопоказаний
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateLinksDto'
      produces:
      - application/json
      responses:
        "200":
          description: Противопоказания обновлены
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Обновление противопоказаний
      tags:
      - Admin
  /admin/medicine/dosage:
    delete:
      consumes:
      - application/json
      description: Удаляет правило дозировки по идентификатору
      parameters:
      - description: Идентификатор правила
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/RemoveDosageRuleDto'
      produces:
      - application/json
      responses:
        "200":
          description: Правило дозировки удалено
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Удаление правила дозировки
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Добавляет правило дозировки к лекарству
      parameters:
      - description: Правило дозировки
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/AddDosageRuleDto'
      produces:
      - application/json
      responses:
        "200":
          description: Правило дозировки добавлено
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Добавление правила дозировки
      tags:
      - Admin
  /admin/medicine/indications:
    patch:
      consumes:
      - application/json
      description: Обновляет список показаний для лекарства
      parameters:
      - description: Список показаний
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/UpdateLinksDto'
      produces:
      - application/json
      responses:
        "200":
          description: Показания обновлены
          schema:
            $ref: '#/definitions/SuccessResDTO'
        "400":
          description: Невалидные входные данные
          schema:
            $ref: '#/definitions/AppError'
        "401":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/AppError'
      security:
      - BearerAuth: []
      summary: Обновление показаний
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by your token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"